Moved 'about' activities
